8.3 Multi Agent LLM Lecture/onboarding_agent/requirements.txt
pyautogen
python-dotenv
openai>=1.0.0
flaml[automl]